+++ src/main/java/ru/netology/exception/NotFoundException.java
@@ -1,22 +1,22 @@
 package ru.netology.exception;
 
 public class NotFoundException extends RuntimeException {
-  public NotFoundException() {
-  }
+    public NotFoundException() {
+    }
 
-  public NotFoundException(String message) {
-    super(message);
-  }
+    public NotFoundException(String message) {
+        super(message);
+    }
 
-  public NotFoundException(String message, Throwable cause) {
-    super(message, cause);
-  }
+    public NotFoundException(String message, Throwable cause) {
+        super(message, cause);
+    }
 
-  public NotFoundException(Throwable cause) {
-    super(cause);
-  }
+    public NotFoundException(Throwable cause) {
+        super(cause);
+    }
 
-  public NotFoundException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
-    super(message, cause, enableSuppression, writableStackTrace);
-  }
+    public NotFoundException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
+        super(message, cause, enableSuppression, writableStackTrace);
+    }
 }
