+++ src/main/java/ru/netology/repository/PostRepository.java
@@ -1,54 +1,18 @@
 package ru.netology.repository;
 
-import ru.netology.exception.NotFoundException;
+import org.springframework.stereotype.Repository;
 import ru.netology.model.Post;
 
-import java.util.*;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.atomic.AtomicLong;
+import java.util.List;
+import java.util.Optional;
 
-// Stub
-public class PostRepository {
 
-    private final Map<Long, Post> repository = new ConcurrentHashMap<>();
-    private final AtomicLong id = new AtomicLong(0);
+public interface PostRepository {
+    List<Post> all();
 
-    public List<Post> all() {
-        if (repository.isEmpty()) {
-            return Collections.emptyList();
-        }
-        return new ArrayList<>(repository.values());
-    }
+    Optional<Post> getById(long id);
 
-    public Optional<Post> getById(long id) {
-        if (repository.containsKey(id)) {
-            return Optional.of(repository.get(id));
-        }
-        return Optional.empty();
-    }
+    Post save(Post post);
 
-    public Post save(Post post) {
-        String content = post.getContent();
-        long idPost = post.getId();
-        //Если от клиента приходит пост с id=0, значит, это создание нового поста.
-        if (idPost == 0) {
-            id.incrementAndGet();
-            repository.put(id.get(), new Post(id.get(), content));
-            return new Post(id.get(), content);
-        }
-        //Если от клиента приходит пост с id !=0, значит, это сохранение (обновление) существующего поста.
-        if (repository.containsKey(idPost)) {
-            Post p = repository.get(idPost);
-            p.setContent(post.getContent());
-            return new Post(idPost, content);
-        }
-        throw new NotFoundException("couldn't save post");
-    }
-
-    public void removeById(long id) {
-        Post post = repository.remove(id);
-        if (post == null) {
-            throw new NotFoundException("The post with the " + id + " was not deleted");
-        }
-    }
+    void removeById(long id);
 }
