+++ src/main/java/ru/netology/service/PostService.java
@@ -1,14 +1,18 @@
 package ru.netology.service;
 
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 import ru.netology.exception.NotFoundException;
 import ru.netology.model.Post;
 import ru.netology.repository.PostRepository;
 
 import java.util.List;
 
+@Service
 public class PostService {
     private final PostRepository repository;
 
+    @Autowired
     public PostService(PostRepository repository) {
 
         this.repository = repository;
