+++ src/main/java/ru/netology/servlet/MainServlet.java
@@ -1,5 +1,7 @@
 package ru.netology.servlet;
 
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 import ru.netology.common.RequestMethods;
 import ru.netology.controller.PostController;
 import ru.netology.repository.PostRepository;
@@ -15,9 +17,8 @@ public class MainServlet extends HttpServlet {
 
     @Override
     public void init() {
-        final var repository = new PostRepository();
-        final var service = new PostService(repository);
-        controller = new PostController(service);
+        final var context = new AnnotationConfigApplicationContext("ru.netology");
+        controller = context.getBean(PostController.class);
     }
 
     @Override
